% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export_files.R
\name{postprocessRasqualResults}
\alias{postprocessRasqualResults}
\title{Helper function for tabixFetchGenes and tabixFetchSNPs}
\usage{
postprocessRasqualResults(rasqual_df)
}
\description{
Helper function for tabixFetchGenes and tabixFetchSNPs
}

