% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sample_offsets.R
\name{Quantile}
\alias{Quantile}
\title{Helper function for rasqualGcCorrection}
\usage{
Quantile(x, k = 20)
}
\description{
Helper function for rasqualGcCorrection
}
\author{
Natsuhiko Kumasaka
}

