% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_covariates.R
\name{randomize}
\alias{randomize}
\title{Helper function for rasqualMakeCovariates}
\usage{
randomize(x, g = NULL)
}
\description{
Helper function for rasqualMakeCovariates
}
\author{
Natsuhiko Kumasaka
}

