% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export_files.R
\name{rasqualMetadataToCovariates}
\alias{rasqualMetadataToCovariates}
\title{Converts a tidy data frame into a data frame that is suitable for rasqual}
\usage{
rasqualMetadataToCovariates(metadata, genotype_id_column = "genotype_id")
}
\arguments{
\item{metadata}{Sample metadata data frame}

\item{genotype_id_column}{Name of the genotype_id column in metadata}
}
\value{
data.frame with genotype ids as row names
}
\description{
Removes genotype_id column from the data frame and makes it row names instead.
}

